requirements.txt
BeautifulSoup4
requests
weibo
pymongo
selenium
html5lib
Pillow
